requirements.txt
openai>=1.60.0
chromadb>=0.5.5
tiktoken>=0.7.0
python-dotenv>=1.0.1
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
